pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["GHMan2021 <statio2009@gmail.com>"]
packages = [
    { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
